Upgrade bracket-lib, use its Point type, fix warnings

+++ src/components/renderable.rs
@@ -5,7 +5,7 @@ use specs_derive::Component;
 #[derive(Component)]
 #[storage(VecStorage)]
 pub struct Renderable {
-    pub glyph: u8,
+    pub glyph: u16,
     pub fg: RGB,
     pub bg: RGB,
 }
