+++ src/lib/rect.rs
@@ -1,12 +1,12 @@
 pub struct Rect {
-    pub x1: u16,
-    pub x2: u16,
-    pub y1: u16,
-    pub y2: u16,
+    pub x1: i32,
+    pub x2: i32,
+    pub y1: i32,
+    pub y2: i32,
 }
 
 impl Rect {
-    pub fn new(x: u16, y: u16, w: u16, h: u16) -> Rect {
+    pub fn new(x: i32, y: i32, w: i32, h: i32) -> Rect {
         Rect {
             x1: x,
             y1: y,
@@ -20,7 +20,7 @@ impl Rect {
         self.x1 <= other.x2 && self.x2 >= other.x1 && self.y1 <= other.y2 && self.y2 >= other.y1
     }
 
-    pub fn center(&self) -> (u16, u16) {
+    pub fn center(&self) -> (i32, i32) {
         ((self.x1 + self.x2) / 2, (self.y1 + self.y2) / 2)
     }
 }
